requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mssql-django==1.4.2
mysqlclient==2.2.4
PyJWT==2.8.0
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
